requirements.txt
amqp==5.1.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
billiard==3.6.4.0
boto3==1.26.54
botocore==1.29.54
Brotli==1.0.9
celery==5.2.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dj-database-url==1.2.0
Django==4.1.5
django-ckeditor==6.5.1
django-google-analytics-app==6.0.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-social-share==2.3.0
django-storages==1.13.2
django-tinymce==3.5.0
gTTS==2.3.1
gunicorn==20.1.0
idna==3.4
jmespath==1.0.1
kombu==5.2.4
Pillow==9.4.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.0
python-http-client==3.3.7
pytz==2022.7.1
requests==2.28.2
s3transfer==0.6.0
sendgrid==6.9.7
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
structlog==22.3.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.3.0
